requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram
python-telegram-bot==20.7

# AI
dspy-ai==2.4.0
google-generativeai==0.3.2

# Notion
notion-client==2.2.1

# Scheduler
apscheduler==3.10.4

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx~=0.25.2
tenacity==8.2.3
pytz==2024.1

# Dev
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
